server/pyproject.toml
[tool.poetry]
name = "flask-server"
version = "0.1.0"
description = "A flask server"
authors = ["Srija Adhya <srijaadhya.11sc.2020@gmail.com>"]
readme = "../README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-httpauth = "^4.8.0"
jwt = "^1.3.1"
pyjwt = "^2.9.0"
flask-sqlalchemy = "^3.1.1"
sqlalchemy-orm = "^1.2.10"
marshmallow-sqlalchemy = "^1.0.0"
psycopg2-binary = "^2.9.6"
alembic = "^1.13.2"


[tool.poetry.scripts]
start = 'app.__main__:main'


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
